chore(psalm): resolve invalid operands

+++ lib/Listener/RegisterTemplateFileCreatorListener.php
@@ -59,7 +59,7 @@ class RegisterTemplateFileCreatorListener implements IEventListener {
 				$odtType->addMimetype('application/vnd.oasis.opendocument.text-template');
 			}
 			$odtType->setIconSvgInline(file_get_contents($appPath . '/img/x-office-document.svg'));
-			$odtType->setRatio(21 / 29.7);
+			$odtType->setRatio((float)21 / 29.7);
 			return $odtType;
 		});
 		$templateManager->registerTemplateFileCreator(function () use ($ooxml, $appPath) {
